servenv: Refactor to use internal package for mux
