+++ go/vt/servenv/status_test.go
@@ -25,6 +25,8 @@ import (
 
 	"github.com/google/safehtml/template"
 	"github.com/stretchr/testify/require"
+
+	"vitess.io/vitess/go/vt/servenv/testutils"
 )
 
 func init() {
@@ -42,7 +44,7 @@ func init() {
 }
 
 func TestStatus(t *testing.T) {
-	server := HTTPTestServer()
+	server := testutils.HTTPTestServer()
 	defer server.Close()
 
 	resp, err := http.Get(server.URL + StatusURLPath())
@@ -67,7 +69,7 @@ func TestStatus(t *testing.T) {
 }
 
 func TestNamedStatus(t *testing.T) {
-	server := HTTPTestServer()
+	server := testutils.HTTPTestServer()
 	defer server.Close()
 
 	name := "test"
